pyproject.toml
[project]
name = "etl-dbf-oracle"
version = "0.1.0"
description = "A comprehensive, configuration-driven ETL pipeline for Oracle databases with automatic schema management"
readme = "README.md"
requires-python = ">=3.13.5"
dependencies = [
    "dbf>=0.99.10",
    "duckdb>=1.3.2",
    "oracledb>=3.2.0",
    "pipdeptree>=2.28.0",
    "polars>=1.31.0",
    "pyarrow>=21.0.0",
    "python-dotenv>=1.1.1",
    "pyyaml>=6.0.2",
    "pandas>=2.0.0",
    "openpyxl>=3.1.5",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project.scripts]
etl-oracle = "main:main"

[tool.hatch.build.targets.wheel]
packages = ["etl_dbf_oracle"]
